Add file entity.py, change class names
